[Misc] QA prep, bug fixes, formatting

+++ weeb/weeb.py
@@ -1,13 +1,15 @@
+import contextlib
 from random import choice
 
 import discord
+from discord import NotFound
 
 from redbot.core import commands, data_manager
 
 
 class Weeb(commands.Cog):
     """Set of weeby commands to show your otaku-ness\n
-    you can also use 'c' as an additional argument for deleting your message
+    you can use 'c' as an additional argument for deleting your message
     Eg: `[p]uwu c`"""
 
     def __init__(self, bot):
@@ -19,35 +21,41 @@ class Weeb(commands.Cog):
         with open(data_manager.bundled_data_path(self) / "xwx.txt", "r", encoding="utf8") as f:
             self.xwx = f.read().splitlines()
 
-    @commands.command(usage="[c]")
+    @commands.command()
     async def uwu(self, ctx, option: str = None):
         """Replies with UwU variant emoticons\n
         `[p]uwu c` - deletes your message"""
         if option == "c":
-            await ctx.message.delete()
+            if ctx.channel.permissions_for(ctx.me).manage_messages:
+                with contextlib.suppress(NotFound):
+                    await ctx.message.delete()
+            else:
+                raise commands.BotMissingPermissions(discord.Permissions(manage_messages=True))
         await ctx.send(choice(self.uwu))
 
-    @commands.command(usage="[c]")
+    @commands.command()
     async def owo(self, ctx, option: str = None):
         """Replies with OwO variant emoticons
         `[p]owo c` - deletes your message"""
         if option == "c":
-            await ctx.message.delete()
+            if ctx.channel.permissions_for(ctx.me).manage_messages:
+                with contextlib.suppress(NotFound):
+                    await ctx.message.delete()
+            else:
+                raise commands.BotMissingPermissions(discord.Permissions(manage_messages=True))
         await ctx.send(choice(self.owo))
 
-    @commands.command(usage="[c]")
+    @commands.command()
     async def xwx(self, ctx, option: str = None):
-        """Replies with flower girl/yandere girl\n
+        """Replies with flower girl/yandere girl
         `[p]xwx c` - deletes your message"""
         if option == "c":
-            await ctx.message.delete()
+            if ctx.channel.permissions_for(ctx.me).manage_messages:
+                with contextlib.suppress(NotFound):
+                    await ctx.message.delete()
+            else:
+                raise commands.BotMissingPermissions(discord.Permissions(manage_messages=True))
         await ctx.send(choice(self.xwx))
-
-    async def cog_command_error(self, ctx, error):
-        if hasattr(error, "original") and isinstance(error.original, discord.errors.Forbidden):
-            await ctx.send('I require the "Manage Messages" permission to execute that command.')
-        else:
-            await self.bot.on_command_error(ctx, error, unhandled_by_cog=True)
 
     async def red_get_data_for_user(self, *, user_id: int):
         # this cog does not store any data
